Move Partners page components to app router
