+++ app/partners/sections/financial/styles.ts
